refactor: Rename "storage" to "database"
In the future, we may also add the function of importing and exporting book files.
"Storage" will be used. The storage here may be confusing.
Considering that Hive is a kind of DB, so modify "storage" to "database".
