+++ src/ui/flutter_app/lib/models/color_settings.dart
@@ -19,7 +19,7 @@ import 'package:flutter/material.dart' show Colors, Color, immutable;
 import 'package:hive_flutter/adapters.dart'
     show HiveField, HiveType, BinaryReader, BinaryWriter, TypeAdapter;
 import 'package:json_annotation/json_annotation.dart';
-import 'package:sanmill/services/storage/adapters/adapters.dart';
+import 'package:sanmill/services/database/adapters/adapters.dart';
 import 'package:sanmill/shared/theme/colors.dart';
 
 part 'color_settings.g.dart';
