+++ src/ui/flutter_app/lib/models/display_settings.dart
@@ -20,8 +20,8 @@ import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:flutter/material.dart' show Locale, immutable;
 import 'package:hive_flutter/adapters.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:sanmill/services/storage/adapters/adapters.dart';
-import 'package:sanmill/services/storage/storage.dart';
+import 'package:sanmill/services/database/adapters/adapters.dart';
+import 'package:sanmill/services/database/database.dart';
 
 part 'display_settings.g.dart';
