+++ src/ui/flutter_app/lib/models/general_settings.dart
@@ -18,7 +18,7 @@ import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:flutter/foundation.dart' show immutable;
 import 'package:hive_flutter/adapters.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:sanmill/services/storage/adapters/adapters.dart';
+import 'package:sanmill/services/database/adapters/adapters.dart';
 
 part 'general_settings.g.dart';
