+++ src/ui/flutter_app/lib/screens/appearance_settings/appearance_settings_page.dart
@@ -19,9 +19,9 @@ import 'package:hive_flutter/hive_flutter.dart' show Box;
 import 'package:sanmill/generated/intl/l10n.dart';
 import 'package:sanmill/models/color_settings.dart';
 import 'package:sanmill/models/display_settings.dart';
+import 'package:sanmill/services/database/database.dart';
 import 'package:sanmill/services/language_info.dart';
 import 'package:sanmill/services/logger.dart';
-import 'package:sanmill/services/storage/storage.dart';
 import 'package:sanmill/shared/custom_drawer/custom_drawer.dart';
 import 'package:sanmill/shared/settings/settings.dart';
 import 'package:sanmill/shared/theme/app_theme.dart';
