+++ src/ui/flutter_app/lib/screens/game_page/game_page.dart
@@ -21,9 +21,9 @@ import 'package:flutter/material.dart';
 import 'package:sanmill/generated/intl/l10n.dart';
 import 'package:sanmill/models/general_settings.dart';
 import 'package:sanmill/screens/general_settings/general_settings_page.dart';
+import 'package:sanmill/services/database/database.dart';
 import 'package:sanmill/services/logger.dart';
 import 'package:sanmill/services/mill/mill.dart';
-import 'package:sanmill/services/storage/storage.dart';
 import 'package:sanmill/shared/custom_drawer/custom_drawer.dart';
 import 'package:sanmill/shared/custom_spacer.dart';
 import 'package:sanmill/shared/game_toolbar/game_toolbar.dart';
