+++ src/ui/flutter_app/lib/screens/general_settings/reset_settings_alert.dart
@@ -24,7 +24,7 @@ class _ResetSettingsAlert extends StatelessWidget {
   Future<void> _restore(BuildContext context) async {
     Navigator.pop(context);
 
-    await DB.resetStorage();
+    await DB.resetDatabase();
   }
 
   @override
