+++ src/ui/flutter_app/lib/screens/help_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:sanmill/generated/intl/l10n.dart';
-import 'package:sanmill/services/storage/storage.dart';
+import 'package:sanmill/services/database/database.dart';
 import 'package:sanmill/shared/custom_drawer/custom_drawer.dart';
 import 'package:sanmill/shared/theme/app_theme.dart';
