+++ src/ui/flutter_app/lib/screens/home.dart
@@ -30,10 +30,10 @@ import 'package:sanmill/screens/game_page/game_page.dart';
 import 'package:sanmill/screens/general_settings/general_settings_page.dart';
 import 'package:sanmill/screens/help_screen.dart';
 import 'package:sanmill/screens/rule_settings/rule_settings_page.dart';
+import 'package:sanmill/services/database/database.dart';
 import 'package:sanmill/services/environment_config.dart';
 import 'package:sanmill/services/logger.dart';
 import 'package:sanmill/services/mill/mill.dart';
-import 'package:sanmill/services/storage/storage.dart';
 import 'package:sanmill/shared/constants.dart';
 import 'package:sanmill/shared/custom_drawer/custom_drawer.dart';
 import 'package:sanmill/shared/privacy_dialog.dart';
