+++ src/ui/flutter_app/lib/services/database/adapters/src/algorithm_adapter.dart
