+++ src/ui/flutter_app/lib/services/database/adapters/src/color_adapter.dart
