+++ src/ui/flutter_app/lib/services/database/adapters/src/painting_style_adapter.dart
