+++ src/ui/flutter_app/lib/shared/string_buffer_helper.dart
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import 'package:sanmill/services/storage/storage.dart';
+import 'package:sanmill/services/database/database.dart';
 
 /// StringBuffer Extension
 ///
