+++ src/ui/flutter_app/pubspec.yaml
@@ -5,7 +5,7 @@ version: 1.1.39+2611
 
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   animated_text_kit: ^4.1.1
@@ -24,7 +24,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  # local storage
+  # local database
   hive: ^2.0.4
   hive_flutter: ^1.1.0
