+++ src/ui/flutter_app/test/game/position_test.dart
@@ -15,11 +15,11 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import 'package:flutter_test/flutter_test.dart';
+import 'package:sanmill/services/database/database.dart';
 import 'package:sanmill/services/mill/mill.dart';
-import 'package:sanmill/services/storage/storage.dart';
 
 import '../helpers/mocks/audios_mock.dart';
-import '../helpers/mocks/storage_mock.dart';
+import '../helpers/mocks/database_mock.dart';
 import '../helpers/test_mills.dart';
 
 void main() {
